+++ scripts/sources/get-sources.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+
+# Assumes that wget and tar are installed!
+# Assumes the working directory is the current directory.
+
+# Get the Apache Sources
+. "./apache-src.sh"
+
+# Get the OpenSSL Sources
+. "./openssl-src.sh"