+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # apache-scripts Changelog
 
+(05/07/2021) 1.0.4:<br/>
+    - Renamed gsrc.sh script to get-sources.sh<br/>
+    - Split out Apache and OpenSSL source download scripts<br/>
+    - Removed ability to pass in a custom 'OUT' directory in the source scripts- the location of downloaded sources will be reliable<br/>
+    - Updated the sources directory structure to split out 'applications'<br/>
+<br/>
 (05/05/2021) 1.0.3:<br/>
     - Updated gsrc.sh script to build sources into a new directory structure<br/>
     - Added Dockerfile and docker-compose file for testing<br/>